# docs/swagger/swagger.yaml
basePath: /api/v3
definitions:
  order-service_internal_model.Address:
    properties:
      city:
        example: 深圳市
        type: string
      district:
        example: 南山区
        type: string
      name:
        example: 张三
        type: string
      phone:
        example: "13800138000"
        type: string
      province:
        example: 广东省
        type: string
      street:
        example: 科技园路1号
        type: string
      zip_code:
        example: "518000"
        type: string
    type: object
  order-service_internal_model.CreateAddressRequest:
    properties:
      city:
        example: 深圳市
        type: string
      district:
        example: 南山区
        type: string
      name:
        example: 张三
        type: string
      phone:
        example: "13800138000"
        type: string
      province:
        example: 广东省
        type: string
      street:
        example: 科技园路1号
        type: string
      zip_code:
        example: "518000"
        type: string
    required:
    - city
    - district
    - name
    - phone
    - province
    - street
    type: object
  order-service_internal_model.CreateOrderItem:
    properties:
      product_id:
        example: 100
        type: integer
      quantity:
        example: 1
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  order-service_internal_model.CreateOrderRequest:
    properties:
      address:
        $ref: '#/definitions/order-service_internal_model.CreateAddressRequest'
      items:
        items:
          $ref: '#/definitions/order-service_internal_model.CreateOrderItem'
        minItems: 1
        type: array
      user_id:
        example: 1
        type: integer
    required:
    - address
    - items
    - user_id
    type: object
  order-service_internal_model.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: Invalid parameters
        type: string
      message:
        example: Bad Request
        type: string
    type: object
  order-service_internal_model.Order:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        example: CNY
        type: string
      id:
        example: ORD-20240101-001
        type: string
      items:
        items:
          $ref: '#/definitions/order-service_internal_model.OrderItem'
        type: array
      paid_at:
        example: "2024-01-01T01:00:00Z"
        type: string
      shipping_address:
        $ref: '#/definitions/order-service_internal_model.Address'
      status:
        enum:
        - pending
        - paid
        - shipped
        - completed
        - cancelled
        example: pending
        type: string
      total_amount:
        example: 199.99
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  order-service_internal_model.OrderItem:
    properties:
      product_id:
        example: 100
        type: integer
      product_name:
        example: iPhone 15 Pro
        type: string
      quantity:
        example: 1
        type: integer
      total_price:
        example: 7999
        type: number
      unit_price:
        example: 7999
        type: number
    type: object
  order-service_internal_model.OrderListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/order-service_internal_model.Order'
        type: array
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
  order-service_internal_model.OrderStatistics:
    properties:
      cancelled_orders:
        example: 50
        type: integer
      completed_orders:
        example: 900
        type: integer
      pending_orders:
        example: 50
        type: integer
      total_amount:
        example: 999999.99
        type: number
      total_orders:
        example: 1000
        type: integer
    type: object
  order-service_internal_model.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
  order-service_internal_model.UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - pending
        - paid
        - shipped
        - completed
        - cancelled
        example: paid
        type: string
    required:
    - status
    type: object
host: localhost:8002
info:
  contact:
    email: support@company.com
    name: API Support
    url: http://www.company.com/support
  description: 订单管理服务 - 提供订单创建、查询、状态管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Order Service API
  version: 3.0.0
paths:
  /health:
    get:
      description: 检查服务是否正常运行
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 健康检查
      tags:
      - System
  /orders:
    get:
      consumes:
      - application/json
      description: 分页获取订单列表，支持按用户ID和状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 订单状态
        enum:
        - pending
        - paid
        - shipped
        - completed
        - cancelled
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/order-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/order-service_internal_model.OrderListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取订单列表
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: 创建新订单
      parameters:
      - description: 订单信息
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order-service_internal_model.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/order-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/order-service_internal_model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建订单
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: 根据订单ID获取订单详细信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/order-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/order-service_internal_model.Order'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取订单详情
      tags:
      - Orders
  /orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: 根据订单ID取消订单
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/order-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/order-service_internal_model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取消订单
      tags:
      - Orders
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: 根据订单ID更新订单状态
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      - description: 订单状态
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/order-service_internal_model.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/order-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/order-service_internal_model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/order-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新订单状态
      tags:
      - Orders
  /orders/statistics:
    get:
      consumes:
      - application/json
      description: 获取订单统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/order-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/order-service_internal_model.OrderStatistics'
              type: object
      security:
      - BearerAuth: []
      summary: 获取订单统计
      tags:
      - Orders
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"